Update database to receive form submissions
Database is now receiving input from form submission.

I've left a lot of commented code there to help demonstrate
the changes I employed.

Data is now getting to mysql, HOWEVER, note that in order to
have the same type of results, you'll need to seed the users
table in the database before trying to use the form.